Refactor BeautifulSoup for scraping, metadata from FMA, limit no downloads.
Now, the tool uses BeautifulSoup for scraping of all download links and meta-
data from FMA. That is, I removed the depency to mutagen, since metadata like
author, song title, album, etc. is available from the HTML pages directly.

I added a new switch to limit the amount of songs to be downloaded. This is
okay for my use case. If one wanted to download all songs from a genre or the
charts, this switch would need to be adapted.

Metadata is stored as a JSON file in the output directory.